Detect hostname when changing password too
To change a password, a user may not be logged in (necessary to change
an expired password). Since only `login` and `login_kerberos` set the
`_current_hostname` attribute, it was possible to call `change_password`
with `_current_hostname` still set to None. This commit wraps the
`change_password` method in the same hostname detection system as
`login` and `login_kerberos`.

It also does some refactoring to avoid code duplication.

Signed-off-by: Aurélien Bompard <aurelien@bompard.org>